refactor: Create, Read 서비스 클래스 분리
