+++ backend/kirikiri/src/main/java/co/kirikiri/controller/GoalRoomController.java
@@ -2,7 +2,8 @@ package co.kirikiri.controller;
 
 import co.kirikiri.common.interceptor.Authenticated;
 import co.kirikiri.common.resolver.MemberIdentifier;
-import co.kirikiri.service.GoalRoomService;
+import co.kirikiri.service.GoalRoomCreateService;
+import co.kirikiri.service.GoalRoomReadService;
 import co.kirikiri.service.dto.CustomPageRequest;
 import co.kirikiri.service.dto.PageResponse;
 import co.kirikiri.service.dto.goalroom.GoalRoomFilterTypeDto;
@@ -28,19 +29,20 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class GoalRoomController {
 
-    private final GoalRoomService goalRoomService;
+    private final GoalRoomCreateService goalRoomCreateService;
+    private final GoalRoomReadService goalRoomReadService;
 
     @PostMapping
     @Authenticated
     public ResponseEntity<Void> create(@RequestBody @Valid final GoalRoomCreateRequest request,
                                        @MemberIdentifier final String identifier) {
-        final Long id = goalRoomService.create(request, identifier);
+        final Long id = goalRoomCreateService.create(request, identifier);
         return ResponseEntity.created(URI.create("/api/goal-rooms/" + id)).build();
     }
 
     @GetMapping("/{goalRoomId}")
     public ResponseEntity<GoalRoomResponse> findGoalRoom(@PathVariable("goalRoomId") final Long goalRoomId) {
-        final GoalRoomResponse goalRoomResponse = goalRoomService.findGoalRoom(goalRoomId);
+        final GoalRoomResponse goalRoomResponse = goalRoomReadService.findGoalRoom(goalRoomId);
         return ResponseEntity.ok(goalRoomResponse);
     }
 
@@ -48,7 +50,7 @@ public class GoalRoomController {
     @GetMapping(value = "/{goalRoomId}", headers = "Authorization")
     public ResponseEntity<GoalRoomCertifiedResponse> findGoalRoom(@MemberIdentifier final String identifier,
                                                                   @PathVariable("goalRoomId") final Long goalRoomId) {
-        final GoalRoomCertifiedResponse goalRoomResponse = goalRoomService.findGoalRoom(identifier, goalRoomId);
+        final GoalRoomCertifiedResponse goalRoomResponse = goalRoomReadService.findGoalRoom(identifier, goalRoomId);
         return ResponseEntity.ok(goalRoomResponse);
     }
 
@@ -57,7 +59,7 @@ public class GoalRoomController {
             @RequestParam(value = "filterCond", required = false) final GoalRoomFilterTypeDto goalRoomFilterTypeDto,
             @ModelAttribute final CustomPageRequest pageRequest
     ) {
-        final PageResponse<GoalRoomForListResponse> goalRoomsPageResponse = goalRoomService.findGoalRoomsByFilterType(
+        final PageResponse<GoalRoomForListResponse> goalRoomsPageResponse = goalRoomReadService.findGoalRoomsByFilterType(
                 goalRoomFilterTypeDto, pageRequest);
         return ResponseEntity.ok(goalRoomsPageResponse);
     }
