+++ backend/kirikiri/src/main/java/co/kirikiri/controller/RoadmapController.java
@@ -2,7 +2,8 @@ package co.kirikiri.controller;
 
 import co.kirikiri.common.interceptor.Authenticated;
 import co.kirikiri.common.resolver.MemberIdentifier;
-import co.kirikiri.service.RoadmapService;
+import co.kirikiri.service.RoadmapCreateService;
+import co.kirikiri.service.RoadmapReadService;
 import co.kirikiri.service.dto.CustomPageRequest;
 import co.kirikiri.service.dto.PageResponse;
 import co.kirikiri.service.dto.roadmap.request.RoadmapFilterTypeRequest;
@@ -28,19 +29,20 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class RoadmapController {
 
-    private final RoadmapService roadmapService;
+    private final RoadmapCreateService roadmapCreateService;
+    private final RoadmapReadService roadmapReadService;
 
     @PostMapping
     @Authenticated
     public ResponseEntity<Void> create(@MemberIdentifier final String identifier,
                                        @RequestBody @Valid final RoadmapSaveRequest request) {
-        final Long roadmapId = roadmapService.create(request, identifier);
+        final Long roadmapId = roadmapCreateService.create(request, identifier);
         return ResponseEntity.created(URI.create("/api/roadmaps/" + roadmapId)).build();
     }
 
     @GetMapping("/{roadmapId}")
-    public ResponseEntity<RoadmapResponse> getRoadmap(@PathVariable final Long roadmapId) {
-        final RoadmapResponse response = roadmapService.findRoadmap(roadmapId);
+    public ResponseEntity<RoadmapResponse> findRoadmap(@PathVariable final Long roadmapId) {
+        final RoadmapResponse response = roadmapReadService.findRoadmap(roadmapId);
         return ResponseEntity.ok(response);
     }
 
@@ -50,14 +52,14 @@ public class RoadmapController {
             @RequestParam(value = "filterCond", required = false) final RoadmapFilterTypeRequest roadmapFilterTypeRequest,
             @ModelAttribute final CustomPageRequest pageRequest
     ) {
-        final PageResponse<RoadmapResponse> roadmapPageResponse = roadmapService.findRoadmapsByFilterType(
+        final PageResponse<RoadmapResponse> roadmapPageResponse = roadmapReadService.findRoadmapsByFilterType(
                 categoryId, roadmapFilterTypeRequest, pageRequest);
         return ResponseEntity.ok(roadmapPageResponse);
     }
 
     @GetMapping("/categories")
     public ResponseEntity<List<RoadmapCategoryResponse>> findAllRoadmapCategories() {
-        final List<RoadmapCategoryResponse> roadmapCategoryResponses = roadmapService.findAllRoadmapCategories();
+        final List<RoadmapCategoryResponse> roadmapCategoryResponses = roadmapReadService.findAllRoadmapCategories();
         return ResponseEntity.ok(roadmapCategoryResponses);
     }
 }
