+++ backend/kirikiri/src/main/java/co/kirikiri/service/RoadmapReadService.java
@@ -1,26 +1,16 @@
 package co.kirikiri.service;
 
-import co.kirikiri.domain.member.Member;
-import co.kirikiri.domain.member.vo.Identifier;
 import co.kirikiri.domain.roadmap.Roadmap;
 import co.kirikiri.domain.roadmap.RoadmapCategory;
 import co.kirikiri.domain.roadmap.RoadmapContent;
-import co.kirikiri.domain.roadmap.RoadmapDifficulty;
-import co.kirikiri.domain.roadmap.RoadmapNode;
-import co.kirikiri.domain.roadmap.RoadmapNodes;
-import co.kirikiri.exception.AuthenticationException;
 import co.kirikiri.exception.NotFoundException;
-import co.kirikiri.persistence.member.MemberRepository;
 import co.kirikiri.persistence.roadmap.RoadmapCategoryRepository;
 import co.kirikiri.persistence.roadmap.RoadmapContentRepository;
 import co.kirikiri.persistence.roadmap.RoadmapRepository;
 import co.kirikiri.persistence.roadmap.dto.RoadmapFilterType;
 import co.kirikiri.service.dto.CustomPageRequest;
 import co.kirikiri.service.dto.PageResponse;
-import co.kirikiri.service.dto.roadmap.RoadmapNodeSaveDto;
-import co.kirikiri.service.dto.roadmap.RoadmapSaveDto;
 import co.kirikiri.service.dto.roadmap.request.RoadmapFilterTypeRequest;
-import co.kirikiri.service.dto.roadmap.request.RoadmapSaveRequest;
 import co.kirikiri.service.dto.roadmap.response.RoadmapCategoryResponse;
 import co.kirikiri.service.dto.roadmap.response.RoadmapResponse;
 import co.kirikiri.service.mapper.RoadmapMapper;
@@ -34,62 +24,11 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
-public class RoadmapService {
+public class RoadmapReadService {
 
     private final RoadmapRepository roadmapRepository;
     private final RoadmapCategoryRepository roadmapCategoryRepository;
-    private final MemberRepository memberRepository;
     private final RoadmapContentRepository roadmapContentRepository;
-
-    @Transactional
-    public Long create(final RoadmapSaveRequest request, final String identifier) {
-        final Member member = findMemberByIdentifier(identifier);
-        final RoadmapCategory roadmapCategory = findRoadmapCategoryById(request.categoryId());
-        final RoadmapSaveDto roadmapSaveDto = RoadmapMapper.convertToRoadmapSaveDto(request);
-        final Roadmap roadmap = createRoadmap(member, roadmapSaveDto, roadmapCategory);
-
-        return roadmapRepository.save(roadmap).getId();
-    }
-
-    private Member findMemberByIdentifier(final String identifier) {
-        return memberRepository.findByIdentifier(new Identifier(identifier))
-                .orElseThrow(() -> new AuthenticationException("존재하지 않는 회원입니다."));
-    }
-
-    private RoadmapCategory findRoadmapCategoryById(final Long categoryId) {
-        return roadmapCategoryRepository.findById(categoryId)
-                .orElseThrow(() -> new NotFoundException("존재하지 않는 카테고리입니다. categoryId = " + categoryId));
-    }
-
-    private Roadmap createRoadmap(final Member member, final RoadmapSaveDto roadmapSaveDto,
-                                  final RoadmapCategory roadmapCategory) {
-        final RoadmapNodes roadmapNodes = makeRoadmapNodes(roadmapSaveDto.roadmapNodes());
-        final RoadmapContent roadmapContent = makeRoadmapContent(roadmapSaveDto, roadmapNodes);
-        final Roadmap roadmap = makeRoadmap(member, roadmapSaveDto, roadmapCategory);
-        roadmap.addContent(roadmapContent);
-        return roadmap;
-    }
-
-    private RoadmapNodes makeRoadmapNodes(final List<RoadmapNodeSaveDto> roadmapNodeSaveDtos) {
-        return new RoadmapNodes(
-                roadmapNodeSaveDtos.stream()
-                        .map(node -> new RoadmapNode(node.title(), node.content()))
-                        .toList()
-        );
-    }
-
-    private RoadmapContent makeRoadmapContent(final RoadmapSaveDto roadmapSaveDto, final RoadmapNodes roadmapNodes) {
-        final RoadmapContent roadmapContent = new RoadmapContent(roadmapSaveDto.content());
-        roadmapContent.addNodes(roadmapNodes);
-        return roadmapContent;
-    }
-
-    private Roadmap makeRoadmap(final Member member, final RoadmapSaveDto roadmapSaveDto,
-                                final RoadmapCategory roadmapCategory) {
-        return new Roadmap(roadmapSaveDto.title(), roadmapSaveDto.introduction(),
-                roadmapSaveDto.requiredPeriod(), RoadmapDifficulty.valueOf(roadmapSaveDto.difficulty().name()), member,
-                roadmapCategory);
-    }
 
     public RoadmapResponse findRoadmap(final Long id) {
         final Roadmap roadmap = findRoadmapById(id);
