+++ backend/kirikiri/src/test/java/co/kirikiri/service/GoalRoomCreateServiceTest.java
@@ -1,0 +1,163 @@
+package co.kirikiri.service;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyLong;
+import static org.mockito.BDDMockito.given;
+
+import co.kirikiri.domain.goalroom.GoalRoom;
+import co.kirikiri.domain.member.EncryptedPassword;
+import co.kirikiri.domain.member.Gender;
+import co.kirikiri.domain.member.Member;
+import co.kirikiri.domain.member.MemberProfile;
+import co.kirikiri.domain.member.vo.Identifier;
+import co.kirikiri.domain.member.vo.Nickname;
+import co.kirikiri.domain.member.vo.Password;
+import co.kirikiri.domain.roadmap.RoadmapContent;
+import co.kirikiri.domain.roadmap.RoadmapNode;
+import co.kirikiri.domain.roadmap.RoadmapNodes;
+import co.kirikiri.exception.BadRequestException;
+import co.kirikiri.exception.NotFoundException;
+import co.kirikiri.persistence.goalroom.GoalRoomRepository;
+import co.kirikiri.persistence.member.MemberRepository;
+import co.kirikiri.persistence.roadmap.RoadmapContentRepository;
+import co.kirikiri.service.dto.goalroom.request.GoalRoomCreateRequest;
+import co.kirikiri.service.dto.goalroom.request.GoalRoomRoadmapNodeRequest;
+import co.kirikiri.service.dto.goalroom.request.GoalRoomTodoRequest;
+import java.time.LocalDate;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+@ExtendWith(MockitoExtension.class)
+public class GoalRoomCreateServiceTest {
+
+    private static final LocalDate TODAY = LocalDate.now();
+    private static final LocalDate TEN_DAY_LATER = TODAY.plusDays(10);
+
+    private static final RoadmapNode ROADMAP_NODE = new RoadmapNode(1L, "title", "content");
+    private static final RoadmapContent ROADMAP_CONTENT = new RoadmapContent(1L, "content");
+    private static final RoadmapNodes ROADMAP_CONTENTS = new RoadmapNodes(new ArrayList<>(List.of(ROADMAP_NODE)));
+
+    private static Member member;
+    private static MemberProfile memberProfile;
+
+    @Mock
+    private GoalRoomRepository goalRoomRepository;
+
+    @Mock
+    private RoadmapContentRepository roadmapContentRepository;
+
+    @Mock
+    private MemberRepository memberRepository;
+
+    @InjectMocks
+    private GoalRoomCreateService goalRoomService;
+
+    @BeforeAll
+    static void setUp() {
+        ROADMAP_CONTENT.addNodes(ROADMAP_CONTENTS);
+        final Identifier identifier = new Identifier("identifier1");
+        final Password password = new Password("password1!");
+        final EncryptedPassword encryptedPassword = new EncryptedPassword(password);
+        final Nickname nickname = new Nickname("nickname");
+        final String phoneNumber = "010-1234-5678";
+        memberProfile = new MemberProfile(Gender.MALE, TODAY, phoneNumber);
+        member = new Member(identifier, encryptedPassword, nickname, memberProfile);
+    }
+
+    @Test
+    void 정상적으로_골룸을_생성한다() {
+        //given
+        final GoalRoomCreateRequest request = new GoalRoomCreateRequest(1L, "name",
+                20, new GoalRoomTodoRequest("content", TODAY, TEN_DAY_LATER),
+                new ArrayList<>(List.of(new GoalRoomRoadmapNodeRequest(1L, 10, TODAY, TEN_DAY_LATER))));
+
+        given(roadmapContentRepository.findById(anyLong()))
+                .willReturn(Optional.of(ROADMAP_CONTENT));
+        given(memberRepository.findByIdentifier(any()))
+                .willReturn(Optional.of(member));
+        given(goalRoomRepository.save(any()))
+                .willReturn(new GoalRoom(1L, null, null, null));
+
+        //when
+        assertDoesNotThrow(() -> goalRoomService.create(request, member.getIdentifier().getValue()));
+    }
+
+    @Test
+    void 골룸_생성_시_존재하지_않은_로드맵_컨텐츠가_들어올때_예외를_던진다() {
+        //given
+        final GoalRoomCreateRequest request = new GoalRoomCreateRequest(1L, "name",
+                20, new GoalRoomTodoRequest("content", TODAY, TEN_DAY_LATER),
+                new ArrayList<>(List.of(new GoalRoomRoadmapNodeRequest(1L, 10, TODAY, TEN_DAY_LATER))));
+
+        given(roadmapContentRepository.findById(anyLong()))
+                .willReturn(Optional.empty());
+
+        //when
+        //then
+        assertThatThrownBy(() -> goalRoomService.create(request, member.getIdentifier().getValue()))
+                .isInstanceOf(NotFoundException.class);
+    }
+
+    @Test
+    void 골룸_생성_시_로드맵_컨텐츠의_노드사이즈와_요청의_노드사이즈가_다를때_예외를_던진다() {
+        //given
+        final List<GoalRoomRoadmapNodeRequest> wrongSizeGoalRoomRoadmapNodeRequest = new ArrayList<>(List.of(
+                new GoalRoomRoadmapNodeRequest(1L, 10, TODAY, TEN_DAY_LATER),
+                new GoalRoomRoadmapNodeRequest(2L, 10, TODAY, TEN_DAY_LATER)));
+        final GoalRoomCreateRequest request = new GoalRoomCreateRequest(1L, "name",
+                20, new GoalRoomTodoRequest("content", TODAY, TEN_DAY_LATER),
+                wrongSizeGoalRoomRoadmapNodeRequest);
+
+        given(roadmapContentRepository.findById(anyLong()))
+                .willReturn(Optional.of(ROADMAP_CONTENT));
+
+        //when
+        //then
+        assertThatThrownBy(() -> goalRoomService.create(request, member.getIdentifier().getValue()))
+                .isInstanceOf(BadRequestException.class);
+    }
+
+    @Test
+    void 골룸_생성_시_로드맵에_존재하지_않는_노드가_요청으로_들어올때_예외를_던진다() {
+        //given
+        final long wrongRoadmapNodId = 2L;
+        final GoalRoomCreateRequest request = new GoalRoomCreateRequest(1L, "name",
+                20, new GoalRoomTodoRequest("content", TODAY, TEN_DAY_LATER),
+                new ArrayList<>(List.of(new GoalRoomRoadmapNodeRequest(wrongRoadmapNodId, 10, TODAY, TEN_DAY_LATER))));
+
+        given(roadmapContentRepository.findById(anyLong()))
+                .willReturn(Optional.of(ROADMAP_CONTENT));
+
+        //when
+        //then
+        assertThatThrownBy(() -> goalRoomService.create(request, member.getIdentifier().getValue()))
+                .isInstanceOf(NotFoundException.class);
+    }
+
+    @Test
+    void 골룸_생성_시_존재하지_않은_회원의_Identifier가_들어올때_예외를_던진다() {
+        //given
+        final GoalRoomCreateRequest request = new GoalRoomCreateRequest(1L, "name",
+                20, new GoalRoomTodoRequest("content", TODAY, TEN_DAY_LATER),
+                new ArrayList<>(List.of(new GoalRoomRoadmapNodeRequest(1L, 10, TODAY, TEN_DAY_LATER))));
+
+        given(roadmapContentRepository.findById(anyLong()))
+                .willReturn(Optional.of(ROADMAP_CONTENT));
+        given(memberRepository.findByIdentifier(any()))
+                .willReturn(Optional.empty());
+
+        //when
+        //then
+        assertThatThrownBy(() -> goalRoomService.create(request, member.getIdentifier().getValue()))
+                .isInstanceOf(NotFoundException.class);
+    }
+}
