+++ backend/kirikiri/src/test/java/co/kirikiri/service/GoalRoomReadServiceTest.java
@@ -2,9 +2,7 @@ package co.kirikiri.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.when;
 
@@ -33,18 +31,12 @@ import co.kirikiri.domain.roadmap.RoadmapNode;
 import co.kirikiri.domain.roadmap.RoadmapNodeImage;
 import co.kirikiri.domain.roadmap.RoadmapNodeImages;
 import co.kirikiri.domain.roadmap.RoadmapNodes;
-import co.kirikiri.exception.BadRequestException;
 import co.kirikiri.exception.NotFoundException;
 import co.kirikiri.persistence.goalroom.GoalRoomPendingMemberRepository;
 import co.kirikiri.persistence.goalroom.GoalRoomRepository;
-import co.kirikiri.persistence.member.MemberRepository;
-import co.kirikiri.persistence.roadmap.RoadmapContentRepository;
 import co.kirikiri.service.dto.CustomPageRequest;
 import co.kirikiri.service.dto.PageResponse;
 import co.kirikiri.service.dto.goalroom.GoalRoomFilterTypeDto;
-import co.kirikiri.service.dto.goalroom.request.GoalRoomCreateRequest;
-import co.kirikiri.service.dto.goalroom.request.GoalRoomRoadmapNodeRequest;
-import co.kirikiri.service.dto.goalroom.request.GoalRoomTodoRequest;
 import co.kirikiri.service.dto.goalroom.response.GoalRoomCertifiedResponse;
 import co.kirikiri.service.dto.goalroom.response.GoalRoomForListResponse;
 import co.kirikiri.service.dto.goalroom.response.GoalRoomNodeResponse;
@@ -52,10 +44,8 @@ import co.kirikiri.service.dto.goalroom.response.GoalRoomResponse;
 import co.kirikiri.service.dto.member.response.MemberResponse;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -65,134 +55,21 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 
 @ExtendWith(MockitoExtension.class)
-class GoalRoomServiceTest {
+class GoalRoomReadServiceTest {
 
     private static final LocalDate TODAY = LocalDate.now();
     private static final LocalDate TEN_DAY_LATER = TODAY.plusDays(10);
     private static final LocalDate TWENTY_DAY_LAYER = TODAY.plusDays(20);
     private static final LocalDate THIRTY_DAY_LATER = TODAY.plusDays(30);
-
-    private static final RoadmapNode ROADMAP_NODE = new RoadmapNode(1L, "title", "content");
-    private static final RoadmapContent ROADMAP_CONTENT = new RoadmapContent(1L, "content");
-    private static final RoadmapNodes ROADMAP_CONTENTS = new RoadmapNodes(new ArrayList<>(List.of(ROADMAP_NODE)));
-
-    private static Member member;
-    private static MemberProfile memberProfile;
-
+    
     @Mock
     private GoalRoomRepository goalRoomRepository;
-
-    @Mock
-    private RoadmapContentRepository roadmapContentRepository;
-
-    @Mock
-    private MemberRepository memberRepository;
 
     @Mock
     private GoalRoomPendingMemberRepository goalRoomPendingMemberRepository;
 
     @InjectMocks
-    private GoalRoomService goalRoomService;
-
-    @BeforeAll
-    static void setUp() {
-        ROADMAP_CONTENT.addNodes(ROADMAP_CONTENTS);
-        final Identifier identifier = new Identifier("identifier1");
-        final Password password = new Password("password1!");
-        final EncryptedPassword encryptedPassword = new EncryptedPassword(password);
-        final Nickname nickname = new Nickname("nickname");
-        final String phoneNumber = "010-1234-5678";
-        memberProfile = new MemberProfile(Gender.MALE, TODAY, phoneNumber);
-        member = new Member(identifier, encryptedPassword, nickname, memberProfile);
-    }
-
-    @Test
-    void 정상적으로_골룸을_생성한다() {
-        //given
-        final GoalRoomCreateRequest request = new GoalRoomCreateRequest(1L, "name",
-                20, new GoalRoomTodoRequest("content", TODAY, TEN_DAY_LATER),
-                new ArrayList<>(List.of(new GoalRoomRoadmapNodeRequest(1L, 10, TODAY, TEN_DAY_LATER))));
-
-        given(roadmapContentRepository.findById(anyLong()))
-                .willReturn(Optional.of(ROADMAP_CONTENT));
-        given(memberRepository.findByIdentifier(any()))
-                .willReturn(Optional.of(member));
-        given(goalRoomRepository.save(any()))
-                .willReturn(new GoalRoom(1L, null, null, null));
-
-        //when
-        assertDoesNotThrow(() -> goalRoomService.create(request, member.getIdentifier().getValue()));
-    }
-
-    @Test
-    void 골룸_생성_시_존재하지_않은_로드맵_컨텐츠가_들어올때_예외를_던진다() {
-        //given
-        final GoalRoomCreateRequest request = new GoalRoomCreateRequest(1L, "name",
-                20, new GoalRoomTodoRequest("content", TODAY, TEN_DAY_LATER),
-                new ArrayList<>(List.of(new GoalRoomRoadmapNodeRequest(1L, 10, TODAY, TEN_DAY_LATER))));
-
-        given(roadmapContentRepository.findById(anyLong()))
-                .willReturn(Optional.empty());
-
-        //when
-        //then
-        assertThatThrownBy(() -> goalRoomService.create(request, member.getIdentifier().getValue()))
-                .isInstanceOf(NotFoundException.class);
-    }
-
-    @Test
-    void 골룸_생성_시_로드맵_컨텐츠의_노드사이즈와_요청의_노드사이즈가_다를때_예외를_던진다() {
-        //given
-        final List<GoalRoomRoadmapNodeRequest> wrongSizeGoalRoomRoadmapNodeRequest = new ArrayList<>(List.of(
-                new GoalRoomRoadmapNodeRequest(1L, 10, TODAY, TEN_DAY_LATER),
-                new GoalRoomRoadmapNodeRequest(2L, 10, TODAY, TEN_DAY_LATER)));
-        final GoalRoomCreateRequest request = new GoalRoomCreateRequest(1L, "name",
-                20, new GoalRoomTodoRequest("content", TODAY, TEN_DAY_LATER),
-                wrongSizeGoalRoomRoadmapNodeRequest);
-
-        given(roadmapContentRepository.findById(anyLong()))
-                .willReturn(Optional.of(ROADMAP_CONTENT));
-
-        //when
-        //then
-        assertThatThrownBy(() -> goalRoomService.create(request, member.getIdentifier().getValue()))
-                .isInstanceOf(BadRequestException.class);
-    }
-
-    @Test
-    void 골룸_생성_시_로드맵에_존재하지_않는_노드가_요청으로_들어올때_예외를_던진다() {
-        //given
-        final long wrongRoadmapNodId = 2L;
-        final GoalRoomCreateRequest request = new GoalRoomCreateRequest(1L, "name",
-                20, new GoalRoomTodoRequest("content", TODAY, TEN_DAY_LATER),
-                new ArrayList<>(List.of(new GoalRoomRoadmapNodeRequest(wrongRoadmapNodId, 10, TODAY, TEN_DAY_LATER))));
-
-        given(roadmapContentRepository.findById(anyLong()))
-                .willReturn(Optional.of(ROADMAP_CONTENT));
-
-        //when
-        //then
-        assertThatThrownBy(() -> goalRoomService.create(request, member.getIdentifier().getValue()))
-                .isInstanceOf(NotFoundException.class);
-    }
-
-    @Test
-    void 골룸_생성_시_존재하지_않은_회원의_Identifier가_들어올때_예외를_던진다() {
-        //given
-        final GoalRoomCreateRequest request = new GoalRoomCreateRequest(1L, "name",
-                20, new GoalRoomTodoRequest("content", TODAY, TEN_DAY_LATER),
-                new ArrayList<>(List.of(new GoalRoomRoadmapNodeRequest(1L, 10, TODAY, TEN_DAY_LATER))));
-
-        given(roadmapContentRepository.findById(anyLong()))
-                .willReturn(Optional.of(ROADMAP_CONTENT));
-        given(memberRepository.findByIdentifier(any()))
-                .willReturn(Optional.empty());
-
-        //when
-        //then
-        assertThatThrownBy(() -> goalRoomService.create(request, member.getIdentifier().getValue()))
-                .isInstanceOf(NotFoundException.class);
-    }
+    private GoalRoomReadService goalRoomService;
 
     @Test
     void 골룸_아이디로_골룸_정보를_조회한다() {
