+++ backend/kirikiri/src/test/java/co/kirikiri/service/RoadmapReadServiceTest.java
@@ -4,7 +4,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyLong;
-import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.when;
 
 import co.kirikiri.domain.member.EncryptedPassword;
@@ -18,19 +17,14 @@ import co.kirikiri.domain.roadmap.Roadmap;
 import co.kirikiri.domain.roadmap.RoadmapCategory;
 import co.kirikiri.domain.roadmap.RoadmapContent;
 import co.kirikiri.domain.roadmap.RoadmapDifficulty;
-import co.kirikiri.exception.AuthenticationException;
 import co.kirikiri.exception.NotFoundException;
-import co.kirikiri.persistence.member.MemberRepository;
 import co.kirikiri.persistence.roadmap.RoadmapCategoryRepository;
 import co.kirikiri.persistence.roadmap.RoadmapContentRepository;
 import co.kirikiri.persistence.roadmap.RoadmapRepository;
 import co.kirikiri.service.dto.CustomPageRequest;
 import co.kirikiri.service.dto.PageResponse;
 import co.kirikiri.service.dto.member.response.MemberResponse;
-import co.kirikiri.service.dto.roadmap.request.RoadmapDifficultyType;
 import co.kirikiri.service.dto.roadmap.request.RoadmapFilterTypeRequest;
-import co.kirikiri.service.dto.roadmap.request.RoadmapNodeSaveRequest;
-import co.kirikiri.service.dto.roadmap.request.RoadmapSaveRequest;
 import co.kirikiri.service.dto.roadmap.response.RoadmapCategoryResponse;
 import co.kirikiri.service.dto.roadmap.response.RoadmapContentResponse;
 import co.kirikiri.service.dto.roadmap.response.RoadmapResponse;
@@ -47,7 +41,7 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 
 @ExtendWith(MockitoExtension.class)
-class RoadmapServiceTest {
+class RoadmapReadServiceTest {
 
     private final Member member = new Member(1L, new Identifier("identifier1"),
             new EncryptedPassword(new Password("password1!")), new Nickname("닉네임"),
@@ -60,13 +54,10 @@ class RoadmapServiceTest {
     private RoadmapCategoryRepository roadmapCategoryRepository;
 
     @Mock
-    private MemberRepository memberRepository;
-
-    @Mock
     private RoadmapContentRepository roadmapContentRepository;
 
     @InjectMocks
-    private RoadmapService roadmapService;
+    private RoadmapReadService roadmapService;
 
     @Test
     void 특정_아이디를_가지는_로드맵_단일_조회시_해당_로드맵의_정보를_반환한다() {
@@ -112,83 +103,6 @@ class RoadmapServiceTest {
 
         //then
         assertThatThrownBy(() -> roadmapService.findRoadmap(1L))
-                .isInstanceOf(NotFoundException.class);
-    }
-
-    @Test
-    void 존재하지_않는_카테고리를_입력하면_예외가_발생한다() {
-        // given
-        final RoadmapSaveRequest request = new RoadmapSaveRequest(10L, "로드맵 제목", "로드맵 소개글", "로드맵 본문",
-                RoadmapDifficultyType.DIFFICULT, 30,
-                List.of(new RoadmapNodeSaveRequest("로드맵 노드1", "로드맵 노드1 설명")));
-
-        given(memberRepository.findByIdentifier(any()))
-                .willReturn(Optional.of(member));
-        given(roadmapCategoryRepository.findById(any()))
-                .willReturn(Optional.empty());
-
-        // expect
-        assertThatThrownBy(() -> roadmapService.create(request, "identifier1"))
-                .isInstanceOf(NotFoundException.class);
-    }
-
-    @Test
-    void 로드맵을_생성한다() {
-        // given
-        final String roadmapTitle = "로드맵 제목";
-        final String roadmapIntroduction = "로드맵 소개글";
-        final String roadmapContent = "로드맵 본문";
-        final RoadmapDifficultyType difficulty = RoadmapDifficultyType.DIFFICULT;
-        final int requiredPeriod = 30;
-        final RoadmapCategory category = new RoadmapCategory(1L, "여가");
-
-        final List<RoadmapNodeSaveRequest> roadmapNodes = List.of(
-                new RoadmapNodeSaveRequest("로드맵 노드1 제목", "로드맵 노드1 설명"));
-        final RoadmapSaveRequest request = new RoadmapSaveRequest(1L, roadmapTitle, roadmapIntroduction, roadmapContent,
-                difficulty, requiredPeriod, roadmapNodes);
-
-        given(roadmapCategoryRepository.findById(any()))
-                .willReturn(Optional.of(category));
-        given(roadmapRepository.save(any()))
-                .willReturn(new Roadmap(1L, roadmapTitle, roadmapIntroduction, requiredPeriod,
-                        RoadmapDifficulty.valueOf(difficulty.name()), member, category));
-        when(memberRepository.findByIdentifier(member.getIdentifier()))
-                .thenReturn(Optional.of(member));
-
-        // expect
-        assertThat(roadmapService.create(request, "identifier1"))
-                .isEqualTo(1L);
-    }
-
-    @Test
-    void 로드맵_생성시_존재하지_않는_회원이면_입력하면_예외가_발생한다() {
-        // given
-        final RoadmapSaveRequest request = new RoadmapSaveRequest(10L, "로드맵 제목", "로드맵 소개글", "로드맵 본문",
-                RoadmapDifficultyType.DIFFICULT, 30,
-                List.of(new RoadmapNodeSaveRequest("로드맵 노드1", "로드맵 노드1 설명")));
-
-        given(memberRepository.findByIdentifier(any()))
-                .willReturn(Optional.empty());
-
-        // expect
-        assertThatThrownBy(() -> roadmapService.create(request, "identifier1"))
-                .isInstanceOf(AuthenticationException.class);
-    }
-
-    @Test
-    void 로드맵_생성시_존재하지_않는_카테고리를_입력하면_예외가_발생한다() {
-        // given
-        final RoadmapSaveRequest request = new RoadmapSaveRequest(10L, "로드맵 제목", "로드맵 소개글", "로드맵 본문",
-                RoadmapDifficultyType.DIFFICULT, 30,
-                List.of(new RoadmapNodeSaveRequest("로드맵 노드1", "로드맵 노드1 설명")));
-
-        given(memberRepository.findByIdentifier(any()))
-                .willReturn(Optional.of(member));
-        given(roadmapCategoryRepository.findById(any()))
-                .willReturn(Optional.empty());
-
-        // expect
-        assertThatThrownBy(() -> roadmapService.create(request, "identifier1"))
                 .isInstanceOf(NotFoundException.class);
     }
